Move 'corner' plot code into plot.py
